# Map: comparer optimization