Copied from comment https://github.com/dotnet/fsharp/pull/10845#issuecomment-756998839 to show the code and run CI

Map/Set always use the default comparison. It should be easy to optimize it... One may think... But I've seen #9348, #513, and related. Such a hopeless state to optimize (inline/devirtualize) the comparison :(

This implementation is probably not 100% compatible with the current behavior. It could be if we apply the logic from #9348 snippet, to exclude records & DUs, if that code is correct.

If we take the current NuGet 5.0 release without any changes to Map as a baseline, are we ready to (`Ratio` column)
* [Improve the map performance with] primitive keys by 3.65x - `getItem`,
* .. `struct T : IComparable<T>` by 4.14x - `getItemIntLike`,
* .. struct records by 95% - `getItemIntRecord`
* .. `string` by 52% - `getItemString`,
* but lose 15% for reference types - `getItemRefLike`, 17% for reference type records - `getItemIntRefRecord`

?

I would say that everything could be wrapped by  `struct T : IComparable<T>` + efficient logic there, and if someone uses ref-types as a key they do care about about performance by definition. But I do understand that regressing existing code by 15/17% may be too big. Yet the tradeoff is so great for primitive types.

What do you think? What's the most important use case for the map/set?

The types in the bench are:

```            
    [<StructuralEquality;CustomComparison>]
    type IntLike =
        struct
           val Value: int
           new(v:int) = {Value = v}
           member x.CompareTo(y:IntLike) = x.Value.CompareTo(y.Value)
        end
        
        interface IComparable<IntLike> with
            member x.CompareTo(y) = x.CompareTo(y)
            
        interface IComparable with
            member x.CompareTo(y) = x.CompareTo(y :?> IntLike) 


    type RefLike =
        val Value: int
        new(v:int) = {Value = v}
        member x.CompareTo(y:RefLike) = x.Value.CompareTo(y.Value)
        
        interface IComparable<RefLike> with
            member x.CompareTo(y) = x.CompareTo(y)
            
        interface IComparable with
            member x.CompareTo(y) = x.CompareTo(y :?> RefLike)
            
    [<Struct>]
    type IntRecord =
          { Value1 : int
            Value2 : int
          }

    type IntRefRecord =
          { Value1 : int
            Value2 : int
          }
```

* After - this PR
* Main50  - PR: https://github.com/dotnet/fsharp/pull/10845
* NuGet50 - current NuGet, which doesn't have any changes to Map as we've found here https://github.com/dotnet/fsharp/pull/10768#issuecomment-754677632

|           Method |     Job | BuildConfiguration | Size |      Mean |     Error |   StdDev | Ratio | RatioSD | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |-------- |------------------- |----- |----------:|----------:|---------:|------:|--------:|-----:|-------:|------:|------:|----------:|
|          getItem |   After |              After |  100 |  20.48 ns |  0.935 ns | 0.243 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|          getItem |  Main50 |             Main50 |  100 |  29.40 ns |  0.728 ns | 0.189 ns |  1.44 |    0.02 |    2 |      - |     - |     - |         - |
|          getItem | NuGet50 |            NuGet50 |  100 |  74.83 ns |  6.092 ns | 0.943 ns |  3.65 |    0.08 |    3 |      - |     - |     - |         - |
|                  |         |                    |      |           |           |          |       |         |      |        |       |       |           |
|   getItemIntLike |   After |              After |  100 |  60.45 ns |  0.460 ns | 0.119 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|   getItemIntLike |  Main50 |             Main50 |  100 | 220.00 ns |  4.159 ns | 0.644 ns |  3.64 |    0.02 |    2 | 0.0450 |     - |     - |     283 B |
|   getItemIntLike | NuGet50 |            NuGet50 |  100 | 250.34 ns |  9.453 ns | 1.463 ns |  4.14 |    0.03 |    3 | 0.0449 |     - |     - |     283 B |
|                  |         |                    |      |           |           |          |       |         |      |        |       |       |           |
|    getItemString |   After |              After |  100 |  76.03 ns |  1.027 ns | 0.267 ns |  1.00 |    0.00 |    2 |      - |     - |     - |         - |
|    getItemString |  Main50 |             Main50 |  100 |  68.22 ns |  0.538 ns | 0.083 ns |  0.90 |    0.00 |    1 |      - |     - |     - |         - |
|    getItemString | NuGet50 |            NuGet50 |  100 | 115.77 ns |  0.626 ns | 0.162 ns |  1.52 |    0.01 |    3 |      - |     - |     - |         - |
|                  |         |                    |      |           |           |          |       |         |      |        |       |       |           |
|   getItemRefLike |   After |              After |  100 | 278.41 ns |  3.569 ns | 0.927 ns |  1.00 |    0.00 |    3 |      - |     - |     - |         - |
|   getItemRefLike |  Main50 |             Main50 |  100 | 200.05 ns |  6.346 ns | 1.648 ns |  0.72 |    0.01 |    1 |      - |     - |     - |         - |
|   getItemRefLike | NuGet50 |            NuGet50 |  100 | 236.40 ns |  3.194 ns | 0.494 ns |  0.85 |    0.00 |    2 |      - |     - |     - |         - |
|                  |         |                    |      |           |           |          |       |         |      |        |       |       |           |
| getItemIntRecord |   After |              After |  100 | 113.75 ns |  5.312 ns | 1.380 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
| getItemIntRecord |  Main50 |             Main50 |  100 | 200.78 ns |  1.451 ns | 0.224 ns |  1.77 |    0.02 |    2 | 0.0449 |     - |     - |     283 B |
| getItemIntRecord | NuGet50 |            NuGet50 |  100 | 221.77 ns |  4.146 ns | 1.077 ns |  1.95 |    0.03 |    3 | 0.0450 |     - |     - |     283 B |
|                     |         |                    |      |           |          |          |       |         |      |        |       |       |           |
| getItemIntRefRecord |   After |              After |  100 | 255.08 ns |  5.493 ns | 1.427 ns |  1.00 |    0.00 |    3 |      - |     - |     - |         - |
| getItemIntRefRecord |  Main50 |             Main50 |  100 | 173.02 ns |  3.398 ns | 0.883 ns |  0.68 |    0.01 |    1 |      - |     - |     - |         - |
| getItemIntRefRecord | NuGet50 |            NuGet50 |  100 | 211.41 ns |  2.625 ns | 0.406 ns |  0.83 |    0.01 |    2 |      - |     - |     - |         - |