@buybackoff 2021-01-09
This is what we could get if using:

* static read-only field that indicates if a type implements `IComparable<T>`, like [this](https://github.com/Spreads/Spreads/blob/c08af8f2d5e5902153b30df590e886d925d9e988/src/Spreads.Core/KeyComparer.cs#L78)
* a custom IL that add `.constained` modified to `callvirt` `CompareTo` when the static field is true (not sure if that is verifiable), like [this](https://github.com/Spreads/Spreads.Native/blob/f0648294046a0fde4e911929073b98a7c49780b6/dotnet/src/Spreads.Native/UnsafeEx.il#L248-L261).


```
    type CompareHelper<'T when 'T : comparison>() =
        static let c = LanguagePrimitives.FastGenericComparer

        [<MethodImpl(MethodImplOptions.AggressiveInlining)>]
        static member Compare(l:'T, r:'T):int =
            if Spreads.KeyComparer<'T>.IsIComparable then // JIT-time constant
                // add .constained before callvirt to IComparable.CompareTo
                Spreads.Native.UnsafeEx.CompareToConstrained(&Unsafe.AsRef(&l),&Unsafe.AsRef(&r))
            else
                // with IsIComparable as constant all other cases use IComparer directly, as before
                c.Compare(l, r)
```

|           Method |     Job | BuildConfiguration | Size |      Mean |    Error |   StdDev | Ratio | RatioSD | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |-------- |------------------- |----- |----------:|---------:|---------:|------:|--------:|-----:|-------:|------:|------:|----------:|
|          getItem |   After |              After |  100 |  19.22 ns | 0.732 ns | 0.190 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|          getItem |  Main50 |             Main50 |  100 |  28.86 ns | 0.168 ns | 0.026 ns |  1.50 |    0.02 |    2 |      - |     - |     - |         - |
|          getItem | NuGet50 |            NuGet50 |  100 |  73.80 ns | 0.585 ns | 0.152 ns |  3.84 |    0.04 |    3 |      - |     - |     - |         - |
|                  |         |                    |      |           |          |          |       |         |      |        |       |       |           |
|   getItemIntLike |   After |              After |  100 |  48.77 ns | 1.409 ns | 0.366 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|   getItemIntLike |  Main50 |             Main50 |  100 | 215.65 ns | 3.189 ns | 0.828 ns |  4.42 |    0.04 |    2 | 0.0448 |     - |     - |     283 B |
|   getItemIntLike | NuGet50 |            NuGet50 |  100 | 248.46 ns | 3.763 ns | 0.582 ns |  5.09 |    0.05 |    3 | 0.0445 |     - |     - |     283 B |
|                  |         |                    |      |           |          |          |       |         |      |        |       |       |           |
|    getItemString |   After |              After |  100 |  77.26 ns | 0.935 ns | 0.243 ns |  1.00 |    0.00 |    2 |      - |     - |     - |         - |
|    getItemString |  Main50 |             Main50 |  100 |  68.54 ns | 1.239 ns | 0.322 ns |  0.89 |    0.01 |    1 |      - |     - |     - |         - |
|    getItemString | NuGet50 |            NuGet50 |  100 | 114.88 ns | 1.200 ns | 0.312 ns |  1.49 |    0.00 |    3 |      - |     - |     - |         - |
|                  |         |                    |      |           |          |          |       |         |      |        |       |       |           |
|   getItemRefLike |   After |              After |  100 | 132.77 ns | 1.419 ns | 0.369 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
|   getItemRefLike |  Main50 |             Main50 |  100 | 200.83 ns | 4.727 ns | 1.228 ns |  1.51 |    0.01 |    2 |      - |     - |     - |         - |
|   getItemRefLike | NuGet50 |            NuGet50 |  100 | 236.52 ns | 7.414 ns | 1.147 ns |  1.78 |    0.01 |    3 |      - |     - |     - |         - |
|                  |         |                    |      |           |          |          |       |         |      |        |       |       |           |
| getItemIntRecord |   After |              After |  100 |  93.29 ns | 0.634 ns | 0.098 ns |  1.00 |    0.00 |    1 |      - |     - |     - |         - |
| getItemIntRecord |  Main50 |             Main50 |  100 | 199.74 ns | 8.888 ns | 1.375 ns |  2.14 |    0.01 |    2 | 0.0444 |     - |     - |     283 B |
| getItemIntRecord | NuGet50 |            NuGet50 |  100 | 222.91 ns | 7.227 ns | 1.118 ns |  2.39 |    0.01 |    3 | 0.0443 |     - |     - |     283 B |